examples/django_async/notif/pyproject.toml
[tool.poetry]
name = "notif"
version = "0.1.0"
description = ""
authors = ["Guillaume Gauvrit <guillaume@gauvr.it>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.2"
uvicorn = "^0.17.3"
dj-blacksmith = "^0.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
